Rewrite ">" and "=" in lisp
When it comes to comparisons there are several we could use:

- >
- <
- >=
- <=
- =

We only implement "<" in golang, the other comparisons can be built
from that.  We also use "eq" which does type&value comparison to build
a numerical equality test.

Minimal cores for the win :)

As a consequence of this we need to load the stdlib before running
the test-cases, so we can use ">", etc, not just the sole "<"
implemented in golang