Add support for "macros".
Macros in our implementation are functions which do not have their
arguments evaluated.

Consider the following code:

        (foo (+ 1 2))

When this is executed the function `foo` will be called with a single
argument 3.  That is because *before* calling the argument any arguments
are evaluated.

If `foo` is defined as a macro that will not happen, the arguments will
be passed literally.  So this does not print `3`, instead it prints the
list `(+ 1 2)`.

        (define foo (macro (x) (print x)))
        (foo (+ 1 2))
        -> "(+ 1 2)"

We've introduced a new type, `macro`, which allows these functions to
be identified from the environment.