Added variadic support.
This pull-request removes "optional" parameters for functions/macros,
in preference to the more standard "&rest" support for swallowing up
a variable number of arguments.

Once this vararg support landed rewriting the "(cond)" special form
as a standard macro became possible - proving it was a good idea.

So:

* Optional parameters, inspired by emacs, prefixed with "&" are removed.
* Variadic arguments are supported.
* Using the variadic parameter, now available, (cond..) was
  reimplemented as a macro, rather than in golang.

As our (cond) macro is now more standard its use is now more standard,
and the needless quote was removed from its use in the fizzbuzz.lisp
example.