Add a benchmark function.
In the following construct the values of A, B, and C, are set sequentially

  (let* (A .. B .. C ..)
   ( .. )

This means that if we record the timestamp at A and C, and make B
a call to a function, we can determine how long it took to execute.

This pull-request closes #16 by adding such a benchmark function,
to report on the time taken to run a couple of operations.  This is
a more general solution to the previous example where we had some
hardcoded measurements.