Added apply-pairs
I wrote this when I was consideirng using lisp to allow (alias..)
to operate on pairs, but it didn't work because (alias) itself
wasn't going to work.

Still iterating over list-items in pairs is useful, so this is
a good demonstration if nothing else.

Also updated documentation strings for apply, & apply-hash for
better accuracy and to add see-also links.  (apply*) doesn't
return a result - unlike map.