Simplify our eval() method.
This commit is the first in a chain that will close #94, by
simplifying our eval method.

Here we've recognized that all of our simple types return
themselves when they're evaluated - so we've created a simple
"IsSimpleType" in our primitive interface.

This makes it easy for simple things to return themselves,
and no longer need to have a special case.