Cache LSP completions.
This will close #111 by calculating possible completions only once,
and then reusing them for the duration of our LSP server.

This works because we only complete things within the standard-library
which won't change mid-execution.

If we ever supported completion of the users' code then this approach
would need to be revisited, but we don't do that and don't want to at
the moment.