Renamed the builtin-signature, again.
So we have an interpreter which has STDIN, and STDOUT.  However
builtin functions don't have access to the interpreter so they
can't use either.

Right now only `(print ..)` needs access, but it is possible to
imagine that other built-in functions might need such access.

To avoid import-cycles we can't type the interpreter instance,
but we can pass an `any` / `interface{}` which has a pointer to
the interpreter.

This is a biggish change, but purely mechanical so I'm confident
enough about it.