Added -debug flag, writing to STDERR.
Since we have the I/O indirection for STDIN and STDOUT it makes sense
to add the same for STDERR.  Of course we don't have anything going
there by default, so I've added a -debug flag to make it useful.

For the moment I'm logging the output of macro-expansion, but it
probably makes sense to log other things instead/as well.  Or remove
it.

Ideally our lisp would be able to write to STDOUT and STDERR, but
then we get into adding support for streams and files.  That might
be next..?