Coerce arguments to functions, in some cases
This is a bit of a sneaky change, but it is useful to convert arguments
to lists in some cases:

* If a function is defined that accepts a single parameter.
* BUT multiple arguments are actually passed to that function
* THEN
  * If the parameter is not variadic
  * And the parameter is untyped, or typed as a list
  * REPLACE THE Arguments with a list containing their entries.

This means:

    (set! and (fn* (xs:list) ... )

Can then be called:

    (and true false true true ..)

Because the arguments will be rewritten as if they were called:

    (and (list true false true true ..)

This is general purpose, and applies to all functions.

I _think_ it's a good idea, but we'll see.